+++ apps/condo/domains/meter/tasks/deleteReadingsOfDeletedMeterWorker.spec.js
