+++ apps/condo/domains/meter/tasks/sendVerificationDateReminderTaskWorker.js
@@ -2,8 +2,6 @@ const dayjs = require('dayjs')
 
 const { featureToggleManager } = require('@open-condo/featureflags/featureToggleManager')
 const { getLogger } = require('@open-condo/keystone/logging')
-const { createCronTask } = require('@open-condo/keystone/tasks')
-
 
 const { SEND_METER_VERIFICATION_DATE_REMINDER_TASK } = require('@condo/domains/common/constants/featureflags')
 
@@ -11,7 +9,10 @@ const { sendVerificationDateReminder } = require('./sendVerificationDateReminder
 
 const logger = getLogger('meter/sendVerificationDateReminder')
 
-const sendVerificationDateReminderTaskFn = async (context = null) => {
+/**
+ * Syncs new and cancelled subscriptions
+ */
+const sendVerificationDateReminderTaskWorker = async (context = null) => {
     const isFeatureEnabled = await featureToggleManager.isFeatureEnabled(context, SEND_METER_VERIFICATION_DATE_REMINDER_TASK)
 
     // Skip sending notifications if feature is disabled on https://growthbook.doma.ai/features
@@ -25,9 +26,7 @@ const sendVerificationDateReminderTaskFn = async (context = null) => {
     await sendVerificationDateReminder({ date: dayjs(), searchWindowDaysShift: 30, daysCount: 30 })
 }
 
-/**
- * Syncs new and cancelled subscriptions
- */
-const sendVerificationDateReminderTask = createCronTask('sendVerificationDateReminder', '0 14 * * 1-5', sendVerificationDateReminderTaskFn)
 
-module.exports = sendVerificationDateReminderTask
+module.exports = {
+    sendVerificationDateReminderTaskWorker,
+}
