+++ apps/condo/domains/meter/tasks/sendVerificationDateReminderTaskWorker.spec.js
@@ -6,22 +6,21 @@ const index = require('@app/condo/index')
 
 const { setFakeClientMode, setAllFeatureFlags } = require('@open-condo/keystone/test.utils')
 
-const sendSubmitMeterReadingsPushNotificationsTask = require('@condo/domains/meter/tasks/sendSubmitMeterReadingsPushNotificationsTask')
+const { sendVerificationDateReminderTaskWorker } = require('./sendVerificationDateReminderTaskWorker')
 
 
-describe('Meter verification notification', () => {
+describe('Meter verification notification task', () => {
     setFakeClientMode(index)
 
     describe('feature flag', () => {
         it('checks for proper result on enabled', async () => {
             setAllFeatureFlags(true)
-            expect(await sendSubmitMeterReadingsPushNotificationsTask.delay.fn()).toBeUndefined()
+            expect(await sendVerificationDateReminderTaskWorker()).toBeUndefined()
         })
 
         it('checks for proper result on disabled', async () => {
             setAllFeatureFlags(false)
-            expect(await sendSubmitMeterReadingsPushNotificationsTask.delay.fn()).toEqual('disabled')
+            expect(await sendVerificationDateReminderTaskWorker()).toEqual('disabled')
         })
     })
-
 })
